# openapi.with-responses.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/books": {
      "post": {
        "operationId": "BooksController_createBook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BookResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      },
      "get": {
        "operationId": "BooksController_getBooks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BookResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/public": {
      "get": {
        "operationId": "BooksController_getPublicBooks",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BookResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "operationId": "BooksController_getBookById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BookResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      },
      "put": {
        "operationId": "BooksController_updateBook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BookResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      },
      "delete": {
        "operationId": "BooksController_deleteBook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "example": "Operation completed successfully"
                            }
                          },
                          "required": [
                            "message"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "403 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "404 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
...
  },
  "info": {
    "title": "FLENVN API",
    "description": "API documentation for the FLENVN backend services Includes inferred response schemas for frontend integration. Successful responses are wrapped by ResponseInterceptor as { success, data, timestamp }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "CreateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "example": "English Basics"
          },
          "description": {
            "type": "string",
            "example": "A beginner guide."
          },
          "author": {
            "type": "string",
            "example": "Bill H"
          },
          "content": {
            "type": "string",
            "example": "Book body content..."
          },
          "coverImage": {
            "type": "string",
            "example": "https://cdn.example.com/cover.jpg"
          },
          "isPublic": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "example": "English Basics"
          },
          "description": {
            "type": "string",
            "example": "Updated description"
          },
          "author": {
            "type": "string",
            "example": "Bill H"
          },
          "content": {
            "type": "string",
            "example": "Updated content"
          },
          "coverImage": {
            "type": "string",
            "example": "https://cdn.example.com/cover.jpg"
          },
          "isPublic": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ApiEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {},
          "message": {
            "type": "string",
            "example": "Optional message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        },
        "required": [
          "success",
          "data",
          "timestamp"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "details": {}
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "method",
          "message"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "nullable": true,
            "example": "linh"
          },
          "avatar": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/avatar.jpg"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "level": {
            "type": "number",
            "example": 1
          },
          "exp": {
            "type": "number",
            "example": 0
          },
          "streak": {
            "type": "number",
            "example": 0
          },
          "lastActive": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "isAdmin": {
            "type": "boolean",
            "example": false
          },
          "booksCount": {
            "type": "number",
            "example": 2
          },
          "totalWordsUsed": {
            "type": "number",
            "example": 1200
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        }
      },
...
      "BookResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "English Stories"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Short reading practice"
          },
          "author": {
            "type": "string",
            "nullable": true,
            "example": "FLENVN"
          },
          "coverImage": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/book.png"
          },
          "content": {
            "type": "string",
            "nullable": true,
            "example": "Once upon a time..."
          },
          "wordCount": {
            "type": "number",
            "example": 450
          },
          "totalCards": {
            "type": "number",
            "example": 12
          },
          "isPublic": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "flashcards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlashcardResponse"
            }
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
      "FlashcardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "partOfSpeech": {
            "type": "string",
            "nullable": true,
            "example": "noun"
          },
          "pronunciation": {
            "type": "string",
            "nullable": true,
            "example": "/ˌserənˈdipəti/"
          },
          "definition": {
            "type": "string",
            "nullable": true,
            "example": "The occurrence of events by chance in a happy way."
          },
          "translation": {
            "type": "string",
            "nullable": true,
            "example": "sự tình cờ may mắn"
          },
          "audioUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/audio.mp3"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/image.jpg"
          },
          "example": {
            "type": "string",
            "nullable": true,
            "example": "Finding that book was pure serendipity."
          },
          "exampleTranslation": {
            "type": "string",
            "nullable": true,
            "example": "Tìm thấy cuốn sách đó là một sự tình cờ may mắn."
          },
          "easeFactor": {
            "type": "number",
            "example": 2.5
          },
          "interval": {
            "type": "number",
            "example": 1
          },
          "repetitions": {
            "type": "number",
            "example": 0
          },
          "nextReviewDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "learning",
              "reviewing",
              "mastered"
            ],
            "example": "new"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "bookId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        }
      },
...
    }
  }
}